# Conflicts:
#	src/main/java/com/jiaruiblog/task/executor/PicExecutor.java
#	src/main/java/com/jiaruiblog/task/executor/TxtExecutor.java

+++ src/main/java/com/jiaruiblog/controller/DocLogController.java
@@ -2,6 +2,7 @@ package com.jiaruiblog.controller;
 
 import com.jiaruiblog.auth.Permission;
 import com.jiaruiblog.auth.PermissionEnum;
+import com.jiaruiblog.common.MessageConstant;
 import com.jiaruiblog.entity.dto.BasePageDTO;
 import com.jiaruiblog.entity.dto.BatchIdDTO;
 import com.jiaruiblog.service.IDocLogService;
@@ -9,6 +10,7 @@ import com.jiaruiblog.util.BaseApiResult;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -59,6 +61,9 @@ public class DocLogController {
     @DeleteMapping("removeLog")
     public BaseApiResult removeLog(@RequestBody @Valid BatchIdDTO batchIdDTO, HttpServletRequest request) {
         List<String> logIds = batchIdDTO.getIds();
+        if (CollectionUtils.isEmpty(logIds)) {
+            return BaseApiResult.error(MessageConstant.PARAMS_ERROR_CODE, MessageConstant.PARAMS_IS_NOT_NULL);
+        }
         return docLogService.deleteDocLogBatch(logIds,(String) request.getAttribute("id"));
     }
